+++ lib/src/app/view/sign_in.dart
@@ -17,13 +17,12 @@
 ///
 ///
 
-import 'package:flutter/material.dart';
-
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-
 import 'package:workingmemory/src/view.dart';
 
 import 'package:workingmemory/src/controller.dart';
+
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+
 
 class SignIn extends StatefulWidget {
   const SignIn({Key key}) : super(key: key);